Simplify physics update for Dynamics2D and Collider2D
