+++ crates/modor_physics/src/components/pipeline.rs
@@ -1,8 +1,7 @@
-use crate::components::collider::ColliderUserData;
-use crate::components::collision_groups::CollisionGroupRegistry;
+use crate::components::collider::{ColliderUpdate, ColliderUserData};
 use crate::components::physics_hook::PhysicsHook;
-use crate::{Collider2D, Collision2D, CollisionGroup, DeltaTime, Dynamics2D};
-use modor::{ComponentSystems, Query, SingleRef};
+use crate::{Collider2D, Collision2D, DeltaTime, Dynamics2D};
+use modor::{Query, SingleRef};
 use rapier2d::dynamics::{
     CCDSolver, ImpulseJointSet, IntegrationParameters, IslandManager, MultibodyJointSet,
     RigidBodyBuilder, RigidBodyHandle, RigidBodySet,
@@ -57,7 +56,7 @@ impl Pipeline2D {
     #[run_after_previous_and(
         component(DeltaTime),
         component(PhysicsHook),
-        action(BodyUpdate),
+        component(Dynamics2D),
         action(ColliderUpdate)
     )]
     fn update(
@@ -195,14 +194,3 @@ pub(crate) struct ColliderHandleRemoval(BodyHandleReset);
 
 #[derive(Action)]
 pub(crate) struct UnsynchronizedHandleDeletion(BodyHandleReset, ColliderHandleRemoval);
-
-#[derive(Action)]
-pub(crate) struct BodyUpdate(UnsynchronizedHandleDeletion);
-
-#[derive(Action)]
-pub(crate) struct ColliderUpdate(
-    UnsynchronizedHandleDeletion,
-    BodyUpdate,
-    <CollisionGroup as ComponentSystems>::Action,
-    <CollisionGroupRegistry as ComponentSystems>::Action,
-);
